requirements.txt
#logic
datetime
pandas
numpy
seaborn
requests
google-cloud-bigquery
google-cloud-storage
plotly
prophet


# API
fastapi
pytz
uvicorn

#deploy
mlflow==2.1.1
#python-dotenv
prefect==2.8.1
#psycopg2-binary


#frontend
streamlit
Pillow
